updating main menu layout to include stats, updating the updateprefs to work correctly.

+++ NicolasCageApp/src/com/example/nicolascageapp/AbcsWithNic.java
@@ -181,7 +181,7 @@ public class AbcsWithNic extends Activity
 		if(curBestScore < playerScore)
 		{
 			Editor ed = prefs.edit();
-			ed.putLong(Stats.ABCSWITHNIC_BEST_SCORE, playerScore);
+			ed.putInt(Stats.ABCSWITHNIC_BEST_SCORE, playerScore);
 			ed.commit();
 		}
